Update GenericForm and move to own folder
